Add an image folder and the second recipe page
Added an image folder then I move the existing image inside, plus I added
a new image for the second recipe, then I created the html for the second
recipe page. After it, I put a link in the index page that goes to the
second recipe page. Into the second recipe page, I put the contents such
as the dish's name, image, description, ingredients, and steps to cook.
In there I also added a link back to the home page. Also, I added one
list item into the "Steps" section of Adobo page that says "Serve" as a
proper ending for cooking the dish.