Add the last recipe page and put all the recipe's link in a list in home
Created the third or last recipe page inside the recipes folder. I added
the corresponding image for the recipe and save inside images folder. I
put the contents related to the recipe such as recipe's name, image,
description, ingredients and steps then added a link back to the index
page. Back to the home page, I put all the links in a list so that they
can be display on separate lines. I also fix some mispelled words in
Tonkatsu page.